Flytt all kode som har med søknadsresultat ut i egen fil (også for tester)
