+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/behandlingsresultat/BehandlingsresultatUtils.kt
@@ -15,7 +15,6 @@ import no.nav.familie.ba.sak.kjerne.beregning.domene.AndelTilkjentYtelse
 import no.nav.familie.ba.sak.kjerne.beregning.domene.AndelTilkjentYtelseMedEndreteUtbetalinger
 import no.nav.familie.ba.sak.kjerne.beregning.domene.YtelseType
 import no.nav.familie.ba.sak.kjerne.endretutbetaling.domene.EndretUtbetalingAndel
-import no.nav.familie.ba.sak.kjerne.endretutbetaling.domene.Årsak
 import no.nav.familie.ba.sak.kjerne.eøs.felles.beregning.tilTidslinje
 import no.nav.familie.ba.sak.kjerne.eøs.kompetanse.domene.Kompetanse
 import no.nav.familie.ba.sak.kjerne.grunnlag.personopplysninger.Person
@@ -39,76 +38,6 @@ import no.nav.fpsak.tidsserie.StandardCombinators
 import java.time.YearMonth
 
 object BehandlingsresultatUtils {
-
-    internal fun utledResultatPåSøknad(
-        forrigeAndeler: List<AndelTilkjentYtelse>,
-        nåværendeAndeler: List<AndelTilkjentYtelse>,
-        nåværendePersonResultater: Set<PersonResultat>,
-        personerFremstiltKravFor: List<Aktør>,
-        endretUtbetalingAndeler: List<EndretUtbetalingAndel>
-    ): Søknadsresultat {
-        val resultaterFraAndeler = utledSøknadResultatFraAndelerTilkjentYtelse(
-            forrigeAndeler = forrigeAndeler,
-            nåværendeAndeler = nåværendeAndeler,
-            personerFremstiltKravFor = personerFremstiltKravFor,
-            endretUtbetalingAndeler = endretUtbetalingAndeler
-        )
-
-        val erEksplisittAvslagPåMinstEnPersonFremstiltKravFor = erEksplisittAvslagPåMinstEnPersonFremstiltKravFor(
-            nåværendePersonResultater = nåværendePersonResultater,
-            personerFremstiltKravFor = personerFremstiltKravFor
-        )
-
-        val alleResultater = (
-            if (erEksplisittAvslagPåMinstEnPersonFremstiltKravFor) {
-                resultaterFraAndeler.plus(Søknadsresultat.AVSLÅTT)
-            } else {
-                resultaterFraAndeler
-            }
-            ).distinct()
-
-        return alleResultater.kombinerSøknadsresultater()
-    }
-
-    internal fun List<Søknadsresultat>.kombinerSøknadsresultater(): Søknadsresultat {
-        val resultaterUtenIngenEndringer = this.filter { it != Søknadsresultat.INGEN_RELEVANTE_ENDRINGER }
-
-        return when {
-            this.isEmpty() -> throw Feil("Klarer ikke utlede søknadsresultat")
-            this.size == 1 -> this.single()
-            resultaterUtenIngenEndringer.size == 1 -> resultaterUtenIngenEndringer.single()
-            resultaterUtenIngenEndringer.size == 2 && resultaterUtenIngenEndringer.containsAll(listOf(Søknadsresultat.INNVILGET, Søknadsresultat.AVSLÅTT)) -> Søknadsresultat.DELVIS_INNVILGET
-            else -> throw Feil("Klarer ikke kombinere søknadsresultater: $this")
-        }
-    }
-
-    private fun erEksplisittAvslagPåMinstEnPersonFremstiltKravFor(
-        nåværendePersonResultater: Set<PersonResultat>,
-        personerFremstiltKravFor: List<Aktør>
-    ): Boolean =
-        nåværendePersonResultater
-            .filter { personerFremstiltKravFor.contains(it.aktør) }
-            .any {
-                it.vilkårResultater.erEksplisittAvslagPåPerson()
-            }
-
-    internal fun utledSøknadResultatFraAndelerTilkjentYtelse(
-        forrigeAndeler: List<AndelTilkjentYtelse>,
-        nåværendeAndeler: List<AndelTilkjentYtelse>,
-        personerFremstiltKravFor: List<Aktør>,
-        endretUtbetalingAndeler: List<EndretUtbetalingAndel>
-    ): List<Søknadsresultat> {
-        val alleSøknadsresultater = personerFremstiltKravFor.flatMap { aktør ->
-            utledSøknadResultatFraAndelerTilkjentYtelsePerPerson(
-                forrigeAndelerForPerson = forrigeAndeler.filter { it.aktør == aktør },
-                nåværendeAndelerForPerson = nåværendeAndeler.filter { it.aktør == aktør },
-                endretUtbetalingAndelerForPerson = endretUtbetalingAndeler.filter { it.person?.aktør == aktør }
-            )
-        }
-
-        return alleSøknadsresultater.distinct()
-    }
-
     private fun validerAtBarePersonerFramstiltKravForHarFåttAvslag(
         personerDetErFramstiltKravFor: List<Aktør>,
         vilkårsvurdering: Vilkårsvurdering
@@ -118,53 +47,6 @@ object BehandlingsresultatUtils {
         if (!personerDetErFramstiltKravFor.containsAll(personerSomHarFåttAvslag)) {
             throw Feil("Det eksisterer personer som har fått avslag men som ikke har blitt søkt for i søknaden!")
         }
-    }
-
-    private fun utledSøknadResultatFraAndelerTilkjentYtelsePerPerson(
-        forrigeAndelerForPerson: List<AndelTilkjentYtelse>,
-        nåværendeAndelerForPerson: List<AndelTilkjentYtelse>,
-        endretUtbetalingAndelerForPerson: List<EndretUtbetalingAndel>
-    ): List<Søknadsresultat> {
-        val forrigeTidslinje = AndelTilkjentYtelseTidslinje(forrigeAndelerForPerson)
-        val nåværendeTidslinje = AndelTilkjentYtelseTidslinje(nåværendeAndelerForPerson)
-        val endretUtbetalingTidslinje = EndretUtbetalingAndelTidslinje(endretUtbetalingAndelerForPerson)
-
-        val resultatTidslinje = nåværendeTidslinje.kombinerMed(forrigeTidslinje, endretUtbetalingTidslinje) { nåværende, forrige, endretUtbetalingAndel ->
-            val forrigeBeløp = forrige?.kalkulertUtbetalingsbeløp
-            val nåværendeBeløp = nåværende?.kalkulertUtbetalingsbeløp
-
-            when {
-                nåværendeBeløp == forrigeBeløp || nåværendeBeløp == null -> Søknadsresultat.INGEN_RELEVANTE_ENDRINGER // Ingen endring eller fjernet en andel
-                nåværendeBeløp > 0 -> Søknadsresultat.INNVILGET // Innvilget beløp som er annerledes enn forrige gang
-                nåværendeBeløp == 0 -> {
-                    val endringsperiodeÅrsak = endretUtbetalingAndel?.årsak
-
-                    when {
-                        nåværende.differanseberegnetPeriodebeløp != null -> Søknadsresultat.INNVILGET
-                        endringsperiodeÅrsak == Årsak.DELT_BOSTED -> Søknadsresultat.INNVILGET
-                        (endringsperiodeÅrsak == Årsak.ALLEREDE_UTBETALT) ||
-                            (endringsperiodeÅrsak == Årsak.ENDRE_MOTTAKER) ||
-                            (endringsperiodeÅrsak == Årsak.ETTERBETALING_3ÅR) -> Søknadsresultat.AVSLÅTT
-                        else -> Søknadsresultat.INGEN_RELEVANTE_ENDRINGER
-                    }
-                }
-                else -> Søknadsresultat.INGEN_RELEVANTE_ENDRINGER
-            }
-        }
-
-        return resultatTidslinje.perioder().mapNotNull { it.innhold }.distinct()
-    }
-
-    private fun Set<VilkårResultat>.erEksplisittAvslagPåPerson(): Boolean {
-        // sjekk om vilkårresultater inneholder eksplisitt avslag på et vilkår
-        return this.any { it.erEksplisittAvslagPåSøknad == true }
-    }
-
-    enum class Søknadsresultat {
-        INNVILGET,
-        AVSLÅTT,
-        DELVIS_INNVILGET,
-        INGEN_RELEVANTE_ENDRINGER
     }
 
     internal enum class Endringsresultat {
